AVX-64: Handle partial reads of HTTP header
Adding HTTPS causes some overhead that exposed a flaw of the
original code that handles the parsing of the HTTP header resonse.
The original code assumed there weren't any partial messages.
This caused the program to fail in parsing the HTTP header when
using TLS for HTTPS connection.

This change handles for reading of partial header content to
solve the above issue.