AVX-64: Send NMEA sentence after HTTP GET message
This change removes the NMEA string from the HTTP GET message.
Instead, the NMEA string is sent after the initial HTTP GET message.

The reason for this is to avoid having the server send back RTCM
data before we are done parsing the HTTP response. If this happens
there will be partial RTCM data along with the HTTP response.